refactor: avoiding initial rerenders and duplicated calls
