+++ src/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from './use-fetch'
-export * from './use-storage'
 export * from './use-weather'
